refactor: extract common modules
