--- frontend/src/metabase/embedding/embedding-hub/components/EmbeddingChecklist/index.ts
@@ -1,1 +1,0 @@
-export { EmbeddingChecklist } from "./EmbeddingChecklist";
