+++ frontend/src/metabase/embedding/embedding-hub/components/EmbeddingHubChecklist/index.ts
@@ -1,0 +1,1 @@
+export { EmbeddingHubChecklist } from "./EmbeddingHubChecklist";
