+++ frontend/src/metabase/embedding/embedding-hub/components/index.ts
@@ -1,3 +1,1 @@
 export { EmbeddingHub } from "./EmbeddingHub";
-export { EmbeddingChecklist } from "./EmbeddingChecklist";
-export { DocsLink } from "./DocsLink";
